Moved outside of DSLS (#29)
- Isaac was correct the render methods at this point time are more
related to the controller for that reason I moved it into the controller
folder as Render module.

This is necessary to keep related modules in the same domain. This
should make things more intuitive and cohesive.

Moves the ControllerActions module into a Render module within the
controller, since this is the only place where is currently being used.

This also consolidates the render specs into the controller specs.

This should make things more obvious and coherent.