Routing Scope and Through parameters to routes.
Issue: (Routing Scopes [ch40] )[https://app.clubhouse.io/amber/story/40/route-scopes]

Scopes are a way to group routes under a common path prefix and scoped
set of plug middleware. We might want to do this for admin
functionality, APIs, and especially for versioned APIs. Let's say we
have user generated reviews on a site, and that those reviews first need
to be approved by an admin. The semantics of these resources are quite
different, and they might not share the same controller. Scopes enable
us to segregate these routes.

Overloaded the router.draw method to receive a valve and through
attributes.

This allows to define scopes for routes and allows for cleaner
definition of routes.