[ENV] Allows for environment settings to be loaded dynamically via a YAML settings file (#412)
* [ENV] Add Update YAML Files For Testing

Adds and update the yaml files to be used for testing switching
environments.

- Adds Development YAML
- Adds Encrypted YAML
- Adds Fake Environment Yaml
- Updates Test yaml

Some spec will fail and this is expected so we can work thru the
failing test

* [ENV] Remove Environment Loader Scripts

Removes the Environment loader script files since we will attempt to
load environment dynamically from an Environment Loader Class

* [ENV] Introduces Env Loader

Adds environment loader class and uses this class to load the
environments

- Removes all settings related method from Configuration and only keeps
server related methods
- Adds Environment Loader Spec
- Adds Environment Loader class

* [SESSION] Fixes Session Specs

Fixes session specs and adds Sever.settings scopes

* [Settings] Rename color to colorize_logging

Currently the Settings.color is not very descriptive. After looking at the way Rails does this decided to change the name colorize_logging which it makes more sense

* [Config] Adds inline documentation to application.cr

Adds definitions to each and one of the settings found in the config/application.cr to dictate and clarify to the user the purpose of the setting and how it should be used.

* [CLI] Use application.cr in Exec command

Application.cr is Amber app main entry point and should be use when wanting to load the Amber Environment.

These changes moves towards the goal of using `require “config/application.cr”` as the main entry point for amber app. Increases consistency and cohesion.

* [Server] Removes Configuration File and Cleanups Server

The configuration file was just storing server class methods. The changes found in this PR moves those methods into the `amber/server/server.cr` file and deletes the `amber/server/configuration.cr`

- Moves `amber` module related methods into `amber.cr`
- Creates `amber/server/server.cr` file

* Parrallelize travis

* [Crystal Format] Applies crystal tool format

* fixup! [Config] Adds inline documentation to application.cr

* [Amber] Cleanup settings and logger

* [CLI] Remvoes curly braces from yamls

* [SPECS] Move Build Spec to Root for Fater Specs