[CORS] Adds W3C Cors Specification Support (#665)
* [CORS] Adds W3C Cors Specification Support

Provides facilities for dealing with CORS requests and responses

- A Plug (Handler) that handles CORS request and responses.

\## How It Works

This handler is compliant with the W3C CORS specification. As per
this specification, It doesn’t put any CORS response headers
in a connection that holds an invalid CORS request. To know what
“invalid” CORS request means, have a look at the
“Validity of CORS requests” section below.

When some options that are not mandatory and have no default value
(such :max_age) at intialization, the relative header will often not be
sent at all. This is compliant with the specification and at the
same time it reduces the size of the response, even if just by a handful
of bytes.

The following is a list of all the CORS response headers supported by
the handler:

- Access-Control-Allow-Origin
- Access-Control-Allow-Methods
- Access-Control-Allow-Headers
- Access-Control-Allow-Credentials
- Access-Control-Expose-Headers
- Access-Control-Max-Age

\## Validity of CORS requests

“Invalid CORS request” can mean that a request doesn’t have an Origin
header (so it’s not a CORS request at all) or that it’s a CORS request but:

  - The Origin request header doesn’t match any of the allowed origins
  - The request is a preflight request but it requests to use a method
    or some headers that are not allowed (via the Access-Control-Request-Method
    and Access-Control-Request-Headers headers)

\## Responding to preflight requests

When the request is a preflight request and is a valid one (valid origin, valid
request method, and valid request headers), The CORS handler directly sends a
response to that request instead of just adding headers to the connection.
To do this, the handler halts the connection and sends a response.

* fixup! [CORS] Adds W3C Cors Specification Support

* fixup! [CORS] Adds W3C Cors Specification Support