Support Procs in respond_with (#774)
* Support Procs in respond_with

When using controller#respond_with all content for each given format
is pre-rendered for the request.  There may be a performance penalty when
rendering large lists which are then discarded and/or there are more than
2 formats.

- change the type of @available_responses hash value to allow Procs
- if the response for a format is a Proc use Proc.call as body content
- only set the response status code if it is the default 200
- set the Location header rather than add the url to the header

Procs are supported for formats in #respond_with so that only the request
format is evaluated and rendered.

* Support Procs in respond_with

When using controller#respond_with all content for each given format
is pre-rendered for the request.  There may be a performance penalty when
rendering large lists which are then discarded and/or there are more than
2 formats.

- change the type of @available_responses hash value to allow Procs
- if the response for a format is a Proc use Proc.call as body content
- only set the response status code if it is the default 200
- set the Location header rather than add the url to the header

Procs are supported for formats in #respond_with so that only the request
format is evaluated and rendered.

* add ProcType alias

* cache the value of response body to avoid multiple Proc.call

* ensure body is a string for set_response

* set response.status_code in individual tests

* refactor body method and issue warning when overwriting response.state_code

* fix indentation issue