Dj/update granite (#957)
* update granite version to 0.14.0

* enable primary field in template

I did this to make it easier for users to be able to modify the primary
key.  This is common when using a UUID or some other type of field.

* move create above update

The create is normally above the update since its matches the crecto and
also ordered that way in rails.  This will allow users to feel more
comfortable using amber.

* add table name to auth generator

* fix tests

* fix tests