Increase reliability of "amber exec" output (#1269)
When the code passed to "amber exec" outputs right before exiting, the
output may not be written completely sometimes. This is due to the 1
millisecond delay in the loop when reading the output.

To fix this we just add a millisecond delay after the code exited, so
that the output loop can finish.